Refactor project to src-based mdl package layout
