+++ src/mdl/backtest/__init__.py
@@ -1,0 +1,6 @@
+"""Backtesting engines and metrics."""
+
+from .engine import BacktestParams, run_backtest, run_backtest_signals
+from .metrics import summarize_metrics
+
+__all__ = ["BacktestParams", "run_backtest", "run_backtest_signals", "summarize_metrics"]
