+++ src/mdl/lab/__init__.py
@@ -1,0 +1,5 @@
+"""Strategy research lab orchestration."""
+
+from .strategy_lab import OBJECTIVES, run_strategy_lab
+
+__all__ = ["OBJECTIVES", "run_strategy_lab"]
